% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{bigscale.atac.pseudo}
\alias{bigscale.atac.pseudo}
\title{bigSCale ATAC Pseudo  (IN DEVELOPMENT, DO NOT USE)}
\usage{
bigscale.atac.pseudo(sce, memory.save = FALSE)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class. The required elements are \code{counts(sce)} and \code{rownames(sce)}. Optionally, you can also fill \code{colData(sce)} with any annotation such as batch, condition: they will be displayed in the plots.}
}
\value{
An sce object storing the markers, pseudotime, cluster and other results. To access the results you can use several S4 methods liste below. Also check the online quick start tutorial over
}
\description{
Pseudotime for ATAC-seq data
}
\examples{
sce=bigscale.atac(sce)

}
