% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{bigscale.convert.h5}
\alias{bigscale.convert.h5}
\title{bigscale.convert.h5}
\usage{
bigscale.convert.h5(input.file, output.file, counts.field, filter.cells = 0)
}
\description{
Converts and .HDF5 format to a .MEX format
}
