% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{bigscale.generate.report}
\alias{bigscale.generate.report}
\title{bigscale.generate.report}
\usage{
bigscale.generate.report(sce, file.name)
}
\arguments{
\item{sce}{The single cell object for which you want to generate the report}

\item{file.name}{Directory and/or name of the excel file containing the formatted report}
}
\description{
bigscale.generate.report
}
\examples{
bigscale.generate.report('report.xlsx')
}
