% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{bigscale.readMM}
\alias{bigscale.readMM}
\title{bigscale.readMM}
\usage{
bigscale.readMM(
  file,
  max.vals,
  skip.vals = 0,
  size.only = FALSE,
  current.condition = NA
)
}
\description{
bigscale.readMM
}
