% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{bigscaleDE}
\alias{bigscaleDE}
\title{Differential Expression}
\usage{
bigscaleDE(object, group1, group2, speed.preset = "slow", cap.ones = FALSE)
}
\arguments{
\item{object}{object of the SingleCellExperiment class.}

\item{group1}{a numeric vector with the indices of cells of group1}

\item{group2}{a numeric vector with the indices of cells of group2}

\item{speed.preset}{by default \code{speed.preset='slow'}. It regulates the speed vs. accuracy in the computation of the marker and differentially expressed genes.
\itemize{
  \item {\bold{slow}} { Reccomended for most datasets, provides best marker accuracy but slowest computational time.} 
  \item {\bold{normal}} {A balance between marker accuracy and computational time. }
  \item {\bold{fast}} {Fastest computational time, if you are in a hurry and you have lots of cell (>15K) you can use this}
}}
}
\description{
Performs DE over two groups of cells
}
\examples{
DE=bigscale.DE(sce,group1=c(1:100),group1=c(101:200))
}
