% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{compare.centrality}
\alias{compare.centrality}
\title{Compare gene centralities}
\usage{
compare.centrality(centralities, names.conditions)
}
\arguments{
\item{centralities}{List of at least two elements. Each element must be the (\code{data.frames}) of the centralities previously calculated by \code{compute.network()}.}

\item{names.conditions}{character of the names of the input networks, same length of the \bold{compute.network()}}
}
\value{
A list with a (\code{data.frame}) for each centrality. In each (\code{data.frame}) the genes are ranked for thier change in centrality.
}
\description{
Works with any given number of networks (N).  The centralities previously calculated with \code{compute.network()} for N networks are given as input.
The script sorts the genes accoring to their change in centrality between the first network (first element of the input list) and the other N-1 networks (the rest of the list)
}
\examples{
check the online tutorial at Github

}
