% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{compute.atac.network}
\alias{compute.atac.network}
\title{ATAC-seq Gene regulatory network}
\usage{
compute.atac.network(expr.data, feature.file, quantile.p = 0.998)
}
\arguments{
\item{expr.data}{matrix of expression counts. Works also with sparse matrices of the \pkg{Matrix} package.}

\item{quantile.p}{only the first \eqn{1 - quantile.p} correlations are used to create the edges of the network. If the networ is too sparse(dense) decrease(increase) \eqn{quantile.p}}

\item{gene.names}{character of gene names, now it supports Gene Symbols or Ensembl, Mouse and Human.}

\item{clustering}{type of clustering and correlations computed to infer the network.
\itemize{
 \item {\bold{recursive}} Best quality at the expenses of computational time. If the dataset is larger than 10-15K cells and is highly heterogeneous this can lead to very long computational times (24-48 hours depending of the hardware).
 \item {\bold{direct}} Best trade-off between quality and computational time. If you want to get a quick output not much dissimilar from the top quality of \bold{recursive} one use this option. Can handle quickly also large datasets (>15-20K cells in 30m-2hours depending on hardware)
 \item {\bold{normal}} To be used if the correlations (the output value \bold{cutoff.p}) detected with either \bold{direct} or \bold{recursive} are too low. At the moment, bigSCale displays a warning if the correlation curoff is lower than 0.8 and suggests to eithe use \bold{normal} clustering or increase the input parameter \bold{quantile.p}
}}

\item{speed.preset}{Used only if  \code{clustering='recursive'} . It regulates the speed vs. accuracy of the Zscores calculations. To have a better network quality it is reccomended to use the default \bold{slow}.
\itemize{
  \item {\bold{slow}} {Highly reccomended, the best network quality but the slowest computational time.} 
  \item {\bold{normal}} {A balance between network quality and computational time. }
  \item {\bold{fast}} {Fastest computational time, worste network quality.}
}}

\item{previous.output}{previous output of \code{compute.network()} can be passed as input to evaluate networks with a different quantile.p without re-running the code. Check the online tutorial at https://github.com/iaconogi/bigSCale2.}
}
\value{
A list with the following items:
\itemize{
\item {\bold{centrality}} {Main output: a Data-frame with the network centrality (Degree,Betweenness,Closeness,PAGErank) for each gene(node) of the network}
\item {\bold{graph}} {The regulatory network in iGraph object}
\item {\bold{correlations}} {All pairwise correlations between genes. The correlation is an average between \emph{Pearson} and \emph{Spearman}. Note that it is stored in single precision format (to save memory space) using the package \pkg{float32}.To make any operation or plot on the correlations first transform it to the standard double precisione by running \code{correlations=dbl(correlations)} }
\item {\bold{cutoff.p}} {The adptive cutoff used to select significant correlations}
\item {\bold{tot.scores}} {The Z-scores over which the correlations are computed. The visually check the correlation between to genes \emph{i} and \emph{j} run  \code{plot(tot.scores[,i],tot.scores[,j])} }
\item {\bold{clusters}} {The clusters in which the cells have been partitioned}
\item {\bold{model}} {Bigscale numerical model of the noise}
}
}
\description{
Infers the gene regulatory network from single cell ATAC-seq data
}
\examples{
out=compute.network(expr.data,gene.names)

}
