% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{compute.distances.icell}
\alias{compute.distances.icell}
\title{Compute iCell distances}
\usage{
compute.distances.icell(
  file.dir,
  max.vals,
  pointer,
  driving.genes,
  N_pct,
  edges,
  q.cutoff.narrow,
  q.cutoff.large,
  chunk.number,
  current.condition,
  pooling.factor,
  verbose,
  neighbours
)
}
\description{
Compute iCell distances
}
