% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{compute.network.model}
\alias{compute.network.model}
\title{Compute network model}
\usage{
compute.network.model(expr.data)
}
\arguments{
\item{expr.data}{matrix of expression counts. Works also with sparse matrices of the \pkg{Matrix} package.}
}
\value{
Model
}
\description{
Compute network model
}
\examples{
out=compute.network.model(expr.data,gene.names)

}
