% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{computeMarkers}
\alias{computeMarkers}
\title{computeMarkers}
\usage{
computeMarkers(object, cap.ones = F, ...)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}

\item{speed.preset}{by default \code{speed.preset='slow'}. It regulates the speed vs. accuracy in the computation of the marker and differentially expressed genes.
\itemize{
  \item {\bold{slow}} { Reccomended for most datasets, provides best marker accuracy but slowest computational time.} 
  \item {\bold{normal}} {A balance between marker accuracy and computational time. }
  \item {\bold{fast}} {Fastest computational time, if you are in a hurry and you have lots of cell (>15K) you can use this}
}}
}
\value{
object of the SingleCellExperiment class with raw data of differential expression safely stored inside in an hidden place.
}
\description{
Calculates markers and differentially expressed genes.
}
\examples{
sce=computeMarkers(sce)
}
