% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{extract.cells}
\alias{extract.cells}
\title{extract.cells}
\usage{
extract.cells(input.mex, output.mex, cells, chuncks = 50000)
}
\arguments{
\item{input.mex}{File name of the output dataset}

\item{cells}{Indices of the cells you want to extract}

\item{chunks}{A techincal parameter, you should not really touch this. It represent how large is the chunk of cells read when sequentially pharsing the input.mex}
}
\value{
Writes to disk the subsetted dataset. Please note that if \code{cells} will be automatially sorted. For example, is \code{cells=c(10,1)}, meaning that you want to extract cells 10 and 1, the oupput matrix will contain, in order, cells 1 and 10.
}
\description{
Works with the .mex files. It subsets to the specified set of cells and write them into the output .mex file.
}
\examples{
check the online tutorial at Github

}
