% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{getDistances}
\alias{getDistances}
\title{getDistances}
\usage{
getDistances(object)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\value{
matrix with all cell to cell distances, recovered from virtual memory.
}
\description{
Retrives the cell to cell distances.
}
\examples{
sce=getDistances(sce)
}
