% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{getSignatures}
\alias{getSignatures}
\title{getSignatures}
\usage{
getSignatures(object)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\value{
A list of co-expressed genes.
}
\description{
Retrives the signatures of co-expressed genes.
For more information check the online tutorial at www.github.com
}
\examples{
Signatures=getSignatures(sce)

}
