% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{homogenize.networks}
\alias{homogenize.networks}
\title{homogenize.networks}
\usage{
homogenize.networks(input.networks)
}
\arguments{
\item{input.networks}{A list with in which each element is the output of a previous compute.network()}

\item{tick}{Pearson correlation steps sued to homogenize the numner of edges. The smallest the larger the computational time but the better the homogenization.}
}
\description{
homogenize.networks
}
\examples{
homogenize.networks(list(output1,output2))
}
