% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{preProcess}
\alias{preProcess}
\title{preProcess}
\usage{
preProcess(object, pipeline = "rna")
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class, pre-processed.
}
\description{
Pre-porocessing of the dataset. Removes genes with all zero values, sets the size factors for normalization and other interal things you do not need to know about.
In the future I will add the possibility to add custom size factors here.
}
\examples{
sce=preProcess(sce)

}
