% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{RecursiveClustering}
\alias{RecursiveClustering}
\title{RecursiveClustering}
\usage{
RecursiveClustering(object, modality = "bigscale", fragment = FALSE)
}
\arguments{
\item{fragment}{describe fragment}

\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class, with cell-cell distances stored inside
}
\description{
Clusters in an accurate recursive way to force immediate identification of cell-subtypes
}
\examples{
sce=RecursiveClustering(sce)

}
