% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{restoreData}
\alias{restoreData}
\title{restoreData}
\usage{
restoreData(object)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class, with the assays \code{normcounts()} and \code{transcounts()} restored.
}
\description{
To be run as the last step of the analysis. Retrives data termporarily stored in  virtual memory and stores if safely in the single cell object.
The data was put in virtual meomory in the first place to reduce RAM memory usage.
}
\examples{
sce=restoreData(sce)

}
