% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{setDCT}
\alias{setDCT}
\title{setDCT}
\usage{
setDCT(object, pca.components = 25)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}

\item{modality}{If \code{classic} uses the default bigSCale algorythm. If \code{alternative}, uses an alternative approach which gives a different clustering/TNSE in case you are not satisfied with the default one.}
}
\value{
object of the SingleCellExperiment class, with cell to cell distances stored inside (in the virtual memory)
}
\description{
computes t-SNE, UMAP and clusters
}
\examples{
sce=setDCT(sce)
}
