% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{setMarkers}
\alias{setMarkers}
\title{setMarkers}
\usage{
setMarkers(object, ...)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}

\item{cutoff}{Z-score cutoff to retain only genes with significant changes of expression. By default \code{cutoff=3}. If you feel you have too many(not enough) markers, decrease its value.}
}
\value{
object of the SingleCellExperiment class with markers and signatures safely stored inside.
}
\description{
Organizes the differentially expressed genes in markers of different levels and signatures
}
\examples{
sce=setMarkers(sce)

}
