% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{setModel}
\alias{setModel}
\title{setModel}
\usage{
setModel(object)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class, with the model stored inside.
}
\description{
Computes the numerical model at the core of all bigSCale2 analysis.
}
\examples{
sce=setModel(sce)

}
