% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{setODgenes}
\alias{setODgenes}
\title{setODgenes}
\usage{
setODgenes(object, ...)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}

\item{min_ODscore}{the treshold (Z-score) used to select highly variable genes. By default \core{min_ODscore=2.33}. Increasing(decreasing) it results in less(more) highly variable genes.}
}
\value{
object of the SingleCellExperiment class, with the highly variable genes stored inside.
}
\description{
Computes the numerical model at the core of all bigSCale2 analysis.
}
\examples{
sce=setODgenes(sce)
sce=setODgenes(sce,min_ODscore=2) # I want to use more highly variable geness so I lower the threshold.

}
