% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{storeNormalized}
\alias{storeNormalized}
\title{storeNormalized}
\usage{
storeNormalized(object, memory.save = TRUE)
}
\arguments{
\item{memory.save}{enables a series of tricks to reduce RAM memory usage. Aware of one case (in linux) in which this option causes irreversible error.}

\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class with normalized counts saved inside (in the virtual memory)
}
\description{
Calculates and stores the normalized counts
}
\examples{
sce=storeNormalized(sce)
}
