% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{storePseudo}
\alias{storePseudo}
\title{storePseudo}
\usage{
storePseudo(object)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class with pasudotime data stored inside.
}
\description{
Computes and stores the pseudotime
}
\examples{
sce=storePseudo(sce)
}
