% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{storeTransformed}
\alias{storeTransformed}
\title{storeTransformed}
\usage{
storeTransformed(object, ...)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class with transformed counts saved inside (in the virtual memory)
}
\description{
Calculates and stores a matrix of transformed counts used by bigSCale to make some plots.
}
\examples{
sce=storeTransformed(sce)
}
