% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{storeTsne}
\alias{storeTsne}
\title{storeTsne}
\usage{
storeTsne(object, ...)
}
\arguments{
\item{...}{passed to \code{Rtsne()}}

\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class with TSNE stored inside
}
\description{
Calculates and stores the TSNE of your dataset. It uses the package \pkg{Rtsne}
}
\examples{
sce=storeTsne(sce)
}
