% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{storeUMAP}
\alias{storeUMAP}
\title{storeUMAP}
\usage{
storeUMAP(object, ...)
}
\arguments{
\item{...}{passed to \code{UMAP()}}

\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class with UMAP stored inside
}
\description{
Calculates and stores the UMAP of your dataset. It uses the package \pkg{UMAP}
}
\examples{
sce=storeUMAP(sce)
}
