% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{toCytoscape}
\alias{toCytoscape}
\title{bigscale.toCytoscape}
\usage{
toCytoscape(G, file.name)
}
\arguments{
\item{G}{The graph in igraph format previoulsy calculated with compute.network}

\item{file.name}{Directory and/or name of the Cytoscape file}
}
\description{
bigscale.toCytoscape
}
\examples{
bigscale.toCytoscape(G,'graph.json')
}
