% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{viewGeneBarPlot}
\alias{viewGeneBarPlot}
\title{viewGeneBarPlot}
\usage{
viewGeneBarPlot(object, gene.list)
}
\arguments{
\item{gene.list}{a character vector with a list of gene names}

\item{sce}{object of the SingleCellExperiment class.}
}
\description{
View 2D reduced dimentions. For the moment only supports TSNE.
}
\examples{
viewGeneBarPlot(sce,c("Arx","NeuroD1","Aqp4")) # colors with clusters
}
