% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{viewGeneViolin}
\alias{viewGeneViolin}
\title{viewGeneViolin}
\usage{
viewGeneViolin(object, gene.name, groups)
}
\arguments{
\item{gene.name}{character name of the gene to plot. Must be same names use in \code{rownames()}}

\item{groups}{optional grouping  \code{factor} to be used in place of clusters}

\item{sce}{object of the SingleCellExperiment class.}
}
\description{
View violin plots
}
\examples{
viewGeneViolin(sce,"Penk")
viewGeneViolin(sce,"Penk",groups=custom.groups) # custom.groups must be a factor variable
}
