% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{viewModel}
\alias{viewModel}
\title{viewModel}
\usage{
viewModel(object)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\description{
Plots the numerical model calculated by bigSCale2. The numerical model is the core for all the analysis performed by bigSCale2.
}
\examples{
viewModel(sce)

}
