% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{viewODgenes}
\alias{viewODgenes}
\title{viewODgenes}
\usage{
viewODgenes(object)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\description{
Visualizes a seris of plots showing how the highly variable genes were selected.
}
\examples{
viewODgenes(sce)
}
