% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{ViewPseudo}
\alias{ViewPseudo}
\title{ViewPseudo}
\usage{
ViewPseudo(object, color.by, transform.in)
}
\arguments{
\item{color.by}{\bold{clusters} to color cell by the cluster of belonging. \bold{pseudo} to color cell by the pseudotime. A \bold{gene name} to color cell by its expression.}

\item{transform.in}{adjustmet of gene expression for better visualuzation. Default is \bold{trim}, you can use \bold{log} to visualize \emph{log2(counts)+1}}

\item{sce}{object of the SingleCellExperiment class.}
}
\description{
View the Pseudotime and colors the cell with different options
}
\examples{
ViewPseudo(sce,'clusters')
ViewPseudo(sce,'pseudo')
ViewPseudo(sce,'Olig1')
}
