% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{viewReduced}
\alias{viewReduced}
\title{viewReduced}
\usage{
viewReduced(
  object,
  color.by = "clusters",
  transform.in = "trim",
  method = "TSNE"
)
}
\arguments{
\item{color.by}{\bold{clusters} to color by clusters, or a gene name to color by gene names ( must be the names used in \code{rownames()}, or a factor to color by custom grouping.}

\item{transform.in}{adjustmet of gene expression for better visualuzation. Default is \bold{trim}, you can use \bold{log} to visualize \emph{log2(counts)+1}}

\item{method}{can only be set to \bold{TSNE} for the moment}

\item{sce}{object of the SingleCellExperiment class.}
}
\description{
View 2D reduced dimentions. For the moment only supports TSNE.
}
\examples{
viewReduced(sce,"clusters") # colors with clusters
viewReduced(sce,"Arx") # colors with Arx gene expression. The gene name must be same as in rownames()

}
