% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{viewSignatures}
\alias{viewSignatures}
\title{viewSignatures}
\usage{
viewSignatures(object, selected.cluster)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class.}
}
\value{
A numeric vector containing the bin limits
}
\description{
Pots and heatmap with dendrogram, clusters, pseudotime, transcriptome complexity, custom user colData and expression values for markers and signtures.
}
\examples{
viewSignatures(sce) #plots all the signatures of correlated markers
viewSignatures(sce,selected.cluster=1) # plots the markers (of all levels) of cluster 1

}
