% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{bigscale.atac.pseudo}
\alias{bigscale.atac.pseudo}
\title{bigSCale ATAC Pseudo  (IN DEVELOPMENT, DO NOT USE)}
\usage{
bigscale.atac.pseudo(sce, memory.save = FALSE)
}
\arguments{
\item{sce}{object of the SingleCellExperiment class. The required elements are \code{counts(sce)} and \code{rownames(sce)}.}
}
\value{
An sce object storing the markers cluster and other results. Check the online quick start tutorial \url{https://github.com/iaconogi/bigSCale2#atac-seq-data} to learn more.
}
\description{
Pseudotime for ATAC-seq data
}
\examples{
sce=bigscale.atac(sce)

}
