% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{setClusters}
\alias{setClusters}
\title{setClusters}
\usage{
setClusters(object, customClust, ...)
}
\arguments{
\item{customClust}{(optional) a numeric vector containg your custom cluster assignment, overrides all previous settings.}

\item{sce}{object of the SingleCellExperiment class.}

\item{method.treshold}{(optional) by default \code{method.treshold=0.5}. Increasing(decreasing) it results in bigSCale2 partitioning in more(less) clusters.}

\item{plot.clusters}{(optional) by default \code{plot.clusters=FALSE}. If \code{plot.clusters=TRUE} plots a dendrogram of the clusters while making the analysis.}

\item{cut.depth}{(optional) by default not used. It overrides the internal decisions of bigSCale2 and forces it to cut the dendrogram at cut.depth (0-100%).}
}
\value{
object of the SingleCellExperiment class with the clusters stored inside.
}
\description{
Calculates the clusters of cell types.
}
\examples{
sce=setClusters(sce)
}
