% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{setClusters}
\alias{setClusters}
\title{setClusters}
\usage{
setClusters(
  object,
  customClust = NA,
  classifier = NA,
  num.classifiers = NA,
  k.f = 100,
  ...
)
}
\arguments{
\item{object}{object of the SingleCellExperiment class.}

\item{customClust}{A numeric vector containg your custom cluster assignment, overrides all previous settings.}

\item{classifier}{New option which allows to cluster the cells according to two or three genes given as input.}

\item{num.classifiers}{How many markers should be used for each group. Check online tutorial for further help https://github.com/iaconogi/bigSCale2#classifier}

\item{k.f}{Inversely proportinal to the number of clusters}

\item{plot.clusters}{By default \code{plot.clusters=FALSE}. If \code{plot.clusters=TRUE} plots a dendrogram of the clusters while making the analysis.}

\item{cut.depth}{By default not used. It overrides the internal decisions of bigSCale2 and forces it to cut the dendrogram at cut.depth (0-100 percent). Works ONLY if using bigSCale modality.}
}
\value{
SingleCellExperiment object with the clusters stored inside.
}
\description{
Calculates the clusters of cell types.
}
\examples{
sce=setClusters(sce)
sce=setClusters(sce,classifier='Cd4','Cd8')
}
