% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellMethods.R
\name{setDistances}
\alias{setDistances}
\title{setDistances}
\usage{
setDistances(object, modality = "pca", pca.components = 25)
}
\arguments{
\item{modality}{If \code{classic} uses the default bigSCale algorythm. If \code{alternative}, uses an alternative approach which gives a different clustering/TNSE in case you are not satisfied with the default one.}

\item{sce}{object of the SingleCellExperiment class.}
}
\value{
object of the SingleCellExperiment class, with cell to cell distances stored inside (in the virtual memory)
}
\description{
Computes cell to cell distances using the highly variable genes.
}
\examples{
sce=setDistances(sce)
}
